+++ packages/gamut/src/Form/__tests__/SelectDropdown-test.tsx
@@ -1,3 +1,10 @@
+import {
+  ArrowChevronDownIcon,
+  CalendarIcon,
+  DataTransferVerticalIcon,
+  EarthIcon,
+  MiniChevronDownIcon,
+} from '@codecademy/gamut-icons';
 import { setupEnzyme } from '@codecademy/gamut-tests';
 
 import { SelectDropdown } from '../SelectDropdown';
@@ -17,6 +24,12 @@ const selectOptionsObject = {
   teal: 'teal',
   orange: 'orange',
 };
+
+const optionsIconsArray = [
+  { label: 'first icon', value: 'one', icon: DataTransferVerticalIcon },
+  { label: 'second icon', value: 'two', icon: CalendarIcon },
+  { label: 'third icon', value: 'three', icon: EarthIcon },
+];
 
 const renderWrapper = setupEnzyme(SelectDropdown, {
   options: selectOptions,
@@ -57,8 +70,7 @@ describe('Select', () => {
       size: 'small',
     });
 
-    const dropdownIconTitle = wrapper.find('DropdownIndicator').text();
-    expect(dropdownIconTitle).toEqual('Mini Chevron Down Icon');
+    expect(wrapper.find(MiniChevronDownIcon)).toBeDefined();
   });
 
   it('renders a medium dropdown when size is "medium"', () => {
@@ -67,8 +79,7 @@ describe('Select', () => {
       size: 'medium',
     });
 
-    const dropdownIconTitle = wrapper.find('DropdownIndicator').text();
-    expect(dropdownIconTitle).toEqual('Arrow Chevron Down Icon');
+    expect(wrapper.find(ArrowChevronDownIcon)).toBeDefined();
   });
 
   it('renders a medium dropdown by default', () => {
@@ -76,8 +87,7 @@ describe('Select', () => {
       options: selectOptionsObject,
     });
 
-    const dropdownIconTitle = wrapper.find('DropdownIndicator').text();
-    expect(dropdownIconTitle).toEqual('Arrow Chevron Down Icon');
+    expect(wrapper.find(ArrowChevronDownIcon)).toBeDefined();
   });
 
   it('renders a dropdown with the correct maxHeight when shownOptionsLimit is specified', () => {
@@ -107,5 +117,20 @@ describe('Select', () => {
 
     const menuList = wrapper.find('MenuList');
     expect(menuList.getDOMNode()).toHaveStyle('max-height : 8rem');
+  });
+
+  it('renders a dropdown with icons', () => {
+    const { wrapper } = renderWrapper({
+      options: optionsIconsArray,
+      size: 'small',
+    });
+
+    wrapper.find('DropdownIndicator').first().simulate('mouseDown', {
+      button: 0,
+    });
+
+    expect(wrapper.find(DataTransferVerticalIcon)).toBeDefined();
+    expect(wrapper.find(CalendarIcon)).toBeDefined();
+    expect(wrapper.find(EarthIcon)).toBeDefined();
   });
 });